+++ libft/ft_memccpy.c
@@ -6,14 +6,14 @@
 /*   By: najlee <najlee@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/22 15:47:09 by najlee            #+#    #+#             */
-/*   Updated: 2020/12/22 18:15:16 by najlee           ###   ########.fr       */
+/*   Updated: 2020/12/23 14:36:38 by najlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 void						*ft_memccpy(void *dest,
-		const void *src, int c, unsigned int n)
+		const void *src, int c, size_t n)
 {
 	unsigned char			*tmp;
 	unsigned const char		*str;
