+++ libft/ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: najlee <najlee@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/22 19:36:19 by najlee            #+#    #+#             */
-/*   Updated: 2020/12/22 19:47:53 by najlee           ###   ########.fr       */
+/*   Updated: 2020/12/23 14:37:06 by najlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void				*ft_memchr(void *ptr, int value, unsigned int num)
+void				*ft_memchr(void *ptr, int value, size_t num)
 {
 	unsigned int	i;
 	unsigned char	*str;
