+++ libft/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: najlee <najlee@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/21 20:24:17 by najlee            #+#    #+#             */
-/*   Updated: 2020/12/22 14:58:43 by najlee           ###   ########.fr       */
+/*   Updated: 2020/12/23 14:37:47 by najlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char *ptr;
